requirements-minimal.txt
langgraph>=0.2.0
langchain>=0.1.0
langchain-core>=0.1.0

pdfplumber>=0.11.0
PyMuPDF>=1.23.0

paddlepaddle
paddleocr
Pillow>=10.0.0
pdf2image>=1.16.0

opencv-python>=4.8.0
numpy>=1.24.0

pandas>=2.0.0

requests>=2.31.0

pydantic>=2.0.0

tqdm>=4.66.0

python-dotenv>=1.0.0

python-dateutil>=2.8.2

